requirements.txt
accelerate==1.10.1
transformers==4.57.0
torch==2.8.0
huggingface_hub[cli]
pyarrow
dotenv
bitsandbytes
sacrebleu
protobuf
sentencepiece
customtkinter
pandas
pillow